Split 'other' into separate files
now to avoid having to have another reorg
in the future and lose git history again
