+++ pypuppetdb/api/pql.py
@@ -15,6 +15,10 @@ log = logging.getLogger(__name__)
 
 
 class PqlAPI(BaseAPI):
+    """
+    This class provides methods that interact with the "pdb/query/v4"
+    PuppetDB API endpoint.
+    """
 
     def _pql(self, pql, request_method='GET'):
         """This method prepares a PQL query to PuppetDB. Actual making
